remove a bunch of state functions
